requirements.txt
requests
click
python-pushover==0.2
